Use another method to optimize performance:hash
I use BKDRHash to implement this work. And I choose 42737 to be the size of
hash table. Thus, when I search the dictionary I look for the hash table
instead of original linked list.
And the spend of time decreases, approaching 0. Besides, cache-miss also decrease from 68.025% to 55.205%. But the time
of append function increases because of addtional calculation of hash function.